+++ lib/commands/create_command/screen_creaters/bloc_grid_screen.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:dcli/dcli.dart';
 import 'package:flutter_bloc_cli/commands/create_command/create_command.dart';
 import 'package:flutter_bloc_cli/data/cli_data_provider.dart';
 import 'package:flutter_bloc_cli/data/constants.dart';
@@ -21,21 +22,21 @@ class BlocGridScreen extends CreateCommand with Generator {
           path: "${Constants.screensDirectoryPath}\\$screenName\\bloc\\${screenName}_bloc.dart".actualPath(),
           content: getBlocFileContent(
             screenName,
-            CreateGenerator.blocFileContent,
+            CreateGenerator.blocGridFileContent,
           ),
         ),
         writeFile(
           path: "${Constants.screensDirectoryPath}\\$screenName\\bloc\\${screenName}_event.dart".actualPath(),
           content: getBlocEventFileContent(
             screenName,
-            CreateGenerator.blocEventFileContent,
+            CreateGenerator.blocEventGridFileContent,
           ),
         ),
         writeFile(
           path: "${Constants.screensDirectoryPath}\\$screenName\\bloc\\${screenName}_state.dart".actualPath(),
           content: getBlocStateFileContent(
             screenName,
-            CreateGenerator.blocStateFileContent,
+            CreateGenerator.blocStateGridFileContent,
           ),
         ),
         writeFile(
@@ -49,6 +50,10 @@ class BlocGridScreen extends CreateCommand with Generator {
         writeFile(
           path: "${Constants.screensDirectoryPath}\\$screenName\\view\\grid_list_item_view.dart".actualPath(),
           content: CreateGenerator.gridItemViewFileContent,
+        ),
+        writeFile(
+          path: "${Constants.screensDirectoryPath}\\$screenName\\view\\loading_view.dart".actualPath(),
+          content: CreateGenerator.gridLoadingViewFileContent,
         ),
         writeFile(
           path: "${Constants.screensDirectoryPath}\\$screenName\\repository\\${screenName}_repository.dart".actualPath(),
@@ -81,7 +86,7 @@ class BlocGridScreen extends CreateCommand with Generator {
           ),
         ]);
       }
-      print("\nSuccess! The $screenName screen has been created successfully.");
+      print(green("\nSuccess! The $screenName screen has been created successfully."));
     }
   }
 }