+++ lib/commands/create_command/screen_creaters/cubit_grid_screen.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:dcli/dcli.dart';
 import 'package:flutter_bloc_cli/commands/create_command/create_command.dart';
 import 'package:flutter_bloc_cli/data/cli_data_provider.dart';
 import 'package:flutter_bloc_cli/data/constants.dart';
@@ -21,14 +22,14 @@ class CubitGridScreen extends CreateCommand with Generator {
           path: "${Constants.screensDirectoryPath}\\$screenName\\cubit\\${screenName}_cubit.dart".actualPath(),
           content: getBlocFileContent(
             screenName,
-            CreateGenerator.cubitFileContent,
+            CreateGenerator.cubitGridFileContent,
           ),
         ),
         writeFile(
           path: "${Constants.screensDirectoryPath}\\$screenName\\cubit\\${screenName}_state.dart".actualPath(),
           content: getBlocStateFileContent(
             screenName,
-            CreateGenerator.cubitStateFileContent,
+            CreateGenerator.cubitStateGridFileContent,
           ),
         ),
         writeFile(
@@ -42,6 +43,10 @@ class CubitGridScreen extends CreateCommand with Generator {
         writeFile(
           path: "${Constants.screensDirectoryPath}\\$screenName\\view\\grid_list_item_view.dart".actualPath(),
           content: CreateGenerator.gridItemViewFileContent,
+        ),
+        writeFile(
+          path: "${Constants.screensDirectoryPath}\\$screenName\\view\\loading_view.dart".actualPath(),
+          content: CreateGenerator.gridLoadingViewFileContent,
         ),
         writeFile(
           path: "${Constants.screensDirectoryPath}\\$screenName\\repository\\${screenName}_repository.dart".actualPath(),
@@ -77,7 +82,7 @@ class CubitGridScreen extends CreateCommand with Generator {
           ),
         ]);
       }
-      print("\nSuccess! The $screenName screen has been created successfully.");
+      print(green("\nSuccess! The $screenName screen has been created successfully."));
     }
   }
 }