+++ CHANGELOG.md
@@ -13,3 +13,9 @@
 ### 1.0.3 - Added screen creation command.
 
 - Added a feature for creating multiple screens at once.
+
+### 1.0.4 - Minor fixes
+
+- Moved the utils directory into the core directory and updated the data directory to core.
+- Error messages are now highlighted in red for better visibility.
+- Success messages are now highlighted in green for better visibility.